+++ CHANGES.md
@@ -13,6 +13,10 @@ All notable changes to this project will be documented in this file.
 
 - Use String::format(..) instead of concatenate with +
 
+- Use common logging instead of jdk logger
+
+- Use Arrays.asList(..) insetead new ArrayList(..) for getFieldOrders() 
+
 
 
 ## [1.3.3] - 2021-10-23