+++ CHANGES.md
@@ -9,7 +9,7 @@ All notable changes to this project will be documented in this file.
 
 
 
-## [Unreleased]
+## [1.3.4] - 2021-11-15
 
 ### Changed